requirements.txt
requests>=3.0.0
beautifulsoup4>=4.12.0
scapy>=2.5.0
colorama>=0.4.6
dnspython>=2.4.0
urllib3>=2.0.0
tqdm>=4.66.0
cryptography>=42.0.0
selenium>=4.20.0
paramiko>=3.4.0
lxml>=5.0.0
pyOpenSSL>=24.0.0
aiodns>=3.0.0
aiohttp>=3.9.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
pynacl>=1.5.0
scikit-learn>=1.4.0
torch>=2.3.0
numpy>=2.0.0